@@ -1,9 +1,12 @@
 package model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.Data;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
-import java.util.Date;
+import java.time.LocalDateTime;
 
+@Data
 @Entity
 @Table(name = "users")
 public class User {
@@ -13,88 +16,30 @@ public class User {
 
     @NotNull
     @Column(name = "is_moderated")
+    @JsonIgnore
     private byte isModerator;
 
     @NotNull
     @Column(name = "reg_time")
-    private Date regTime;
+    @JsonIgnore
+    private LocalDateTime regTime;
 
     @NotNull
     private String name;
 
     @NotNull
+    @JsonIgnore
     private String email;
 
     @NotNull
+    @JsonIgnore
     private String password;
 
     @NotNull
+    @JsonIgnore
     private String code;
 
     @NotNull
+    @JsonIgnore
     private String photo;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public byte getIsModerator() {
-        return isModerator;
-    }
-
-    public void setIsModerator(byte isModerator) {
-        this.isModerator = isModerator;
-    }
-
-    public Date getRegTime() {
-        return regTime;
-    }
-
-    public void setRegTime(Date regTime) {
-        this.regTime = regTime;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getPhoto() {
-        return photo;
-    }
-
-    public void setPhoto(String photo) {
-        this.photo = photo;
-    }
 }
